+++ core-service/src/main/resources/db/migration/V1__create_product_table.sql
@@ -16,20 +16,20 @@ INSERT INTO category (title) VALUES
 CREATE TABLE products (
                          id bigserial primary key,
                          title varchar not null,
-                         price int not null,
+                         price numeric(8,2) not null,
                          category_id bigint references category (id),
                          created_at timestamp default current_timestamp,
                          updated_at timestamp default current_timestamp
 );
 
 INSERT INTO products (title, price) VALUES
-                                       ('Product 1', 100),
-                                       ('Product 2', 101),
-                                       ('Product 3', 102),
-                                       ('Product 4', 103),
-                                       ('Product 5', 104),
-                                       ('Product 6', 105),
-                                       ('Product 7', 106),
-                                       ('Product 8', 107),
-                                       ('Product 9', 108),
-                                       ('Product 10', 109);
+                                       ('Product 1', 100.00),
+                                       ('Product 2', 101.00),
+                                       ('Product 3', 102.00),
+                                       ('Product 4', 103.00),
+                                       ('Product 5', 104.00),
+                                       ('Product 6', 105.00),
+                                       ('Product 7', 106.00),
+                                       ('Product 8', 107.00),
+                                       ('Product 9', 108.00),
+                                       ('Product 10', 109.00);
