+++ core-service/src/test/java/com/geekbrains/ru/gb_rest_angular/core/test/ProductServiceTest.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -29,17 +30,17 @@ public class ProductServiceTest {
     @Test
     public void getAllProductTest(){
         List<Product> products = new ArrayList<>();
-        products.add(new Product(69L,"Milk",60));
-        products.add(new Product(86L,"Chocolate", 120));
-        products.add(new Product(752L,"Apple",87));
+        products.add(new Product(69L,"Milk", BigDecimal.valueOf(60)));
+        products.add(new Product(86L,"Chocolate", BigDecimal.valueOf(120)));
+        products.add(new Product(752L,"Apple",BigDecimal.valueOf(87)));
 
         Mockito.doReturn(products).when(productRepository).findAll();
 
         List<Product> productList = productService.getAllProduct();
 
         Assertions.assertEquals(productList.size(),3);
-        Assertions.assertTrue(productList.contains(new Product(752L,"Apple",87)));
-        Assertions.assertEquals(productList.stream().filter(p->p.getId().equals(69L)).findFirst(), Optional.of(new Product(69L,"Milk",60)));
+        Assertions.assertTrue(productList.contains(new Product(752L,"Apple",BigDecimal.valueOf(87))));
+        Assertions.assertEquals(productList.stream().filter(p->p.getId().equals(69L)).findFirst(), Optional.of(new Product(69L,"Milk",BigDecimal.valueOf(60))));
         Assertions.assertArrayEquals(productList.toArray(),products.toArray());
 
         Mockito.verify(productRepository,Mockito.times(1)).findAll();
@@ -47,11 +48,11 @@ public class ProductServiceTest {
 
     @Test
     public void findProductByIdTest(){
-        Product product = new Product(1683L,"Cheese",650);
+        Product product = new Product(1683L,"Cheese",BigDecimal.valueOf(650));
 
         Mockito.doReturn(Optional.of(product)).when(productRepository).findById(1683L);
         Mockito.doThrow(new ResourceNotFoundException("Product not found")).when(productRepository).findById(1258L);
-        Assertions.assertEquals(new Product(1683L,"Cheese",650),productService.findProductById(1683L).get());
+        Assertions.assertEquals(new Product(1683L,"Cheese",BigDecimal.valueOf(650)),productService.findProductById(1683L).get());
         //Assertions.assertThrows(ResourceNotFoundException, new ResourceNotFoundException("Product not found"));
     }
 }
